Refactor to support primary use case
Some refactoring to better support the primary use case (measuring app
availability).

* Simplify main test loop and focus on synchronous execution
* Discover route URL via configurable named route
* Clean up logging and error handling
* Remove old prototyping code
* Remove manifest stuff for installation (this is in Ansible)
* Update docs

This commit also temporarily removes setup for diagnostic checks of intermediate
steps; diagnostics can be carefully reintroduced in a followup commit.

The intent of this commit is to establish a simplified/clear foundation for the
necessary metrics collection.